♻️ refactor: decouple login logic from login page
